Adding doc

+++ mainScript.pl
@@ -20,6 +20,9 @@ my $fakeRun       = 0;
 #    }
 #  }
 #}
+
+#ugly hack due to the fact that we need to wait for processes writing to file descriptor
+#rather than actual files
 sub findAllPIDs{
  my ($firstPID) = @_;
  my %hashPID;
@@ -72,7 +75,7 @@ sub systemBash {
  my $cmd = shift;
  my @args = ( "bash", "-c", $cmd );
 
- print "running ".$cmd."\n";
+ warn "running ".$cmd."\n";
  if(!$fakeRun){
    my $pid=fork();
    if($pid == 0){
@@ -102,6 +105,7 @@ my $genomicFN     = "none";
 my $inputFile     = "none";
 
 my $allowedMismatches=1;
+my $onlyCreateDatabase=0;
 
 my %opts;
 
@@ -127,13 +131,16 @@ my $usage=
   "$0 [options]\n\n".
   "Main script to automatically build the database, parse and search for the tags\n\n".
   "Database building options\n".
+  "(by default at most 1 mismatch in the recognition site)\n".
   "\t-l Length of sequence for database             (Default:".$length.")\n".
-  "\t-n Allowed mismatches in restriction site      (Default:".$mismatchesRestSize.")\n".
+  #  "\t-n Allowed mismatches in restriction site      (Default:".$mismatchesRestSize.")\n".
   "\t-r Restriction sequence to use e.g. GATC       (Default:".$restSequence.")\n".
   "\t-o Directory to store the database files       (Default:".$outdir.")\n".
   "\t-s Path to the RefSeq database  (fasta format) (Default:".$refseq.")\n".
   "\t-u Path to the Unigene database (fasta format) (Default:".$unigene.")\n".
   "\t-g Path to the Genomic database (fasta format) (Default:".$genomic.")\n".
+  "\t-d Only build the database\n".
+
   "\nTag searching options\n".
   "\t-m Allowed mismatches during search                          (Default:".$allowedMismatches.")\n".
   "\t-i Path to the input file containing the tags (fastq format) (Default:".$inputFile.")\n".
@@ -146,7 +153,12 @@ if($#ARGV == -1 || ($#ARGV==0 && ($ARGV[0] eq "-h" || $ARGV[0] eq "--help" ))){
   die $usage;
 }
 
-getopts('l:n:r:o:s:u:g:m:i:f:', \%opts);
+getopts('l:n:r:o:s:u:g:m:i:fd', \%opts);
+
+if(exists $opts{'d'}){
+  $onlyCreateDatabase=1;
+}
+
 
 
 if(exists $opts{'i'}){
@@ -155,7 +167,9 @@ if(exists $opts{'i'}){
     die "Error, input file = ".$inputFile." does not exist\n";
   }
 }else{
-  die "-i option must be specified\n";
+  if($onlyCreateDatabase != 1){
+    die "-i option must be specified\n";
+  }
 }
 
 
@@ -223,19 +237,22 @@ if(exists $opts{'l'}){
 
 }
 
-if(exists $opts{'n'}){
-  $mismatchesRestSize=$opts{'n'};
-  if($mismatchesRestSize != /^\d+$/){
-    die "-n parameter must be numerical";
-  }
-}else{
-
-}
+#if(exists $opts{'n'}){
+#  $mismatchesRestSize=$opts{'n'};
+#  if($mismatchesRestSize != /^\d+$/){
+#    die "-n parameter must be numerical";
+#  }
+#}else{
+#}
 
 
 if(exists $opts{'m'}){
+  if($onlyCreateDatabase == 1){
+    die "Do not specify both -m and -d";
+  }
+
   $allowedMismatches=$opts{'m'};
-  if($mismatchesRestSize != /^\d+$/){
+  if($allowedMismatches != /^\d+$/){
     die "-n parameter must be numerical";
   }
 }else{
@@ -250,12 +267,12 @@ if(exists $opts{'r'}){
 }
 
 
-
 if(exists $opts{'f'}){
   $fakeRun=1
 }
 
-print "Summary of options:\n".
+
+warn "Summary of options:\n".
   "Length of sequence for database        :\t".$length."\n".
   "Allowed mismatches in restriction site :\t".$mismatchesRestSize."\n".
   "Restriction sequence to use            :\t".$restSequence."\n".
@@ -270,17 +287,18 @@ print "Summary of options:\n".
 
 
 
+
 if ( (-e $outdir.$refseqFN.".".$restSequence.".l".$length.".mm0.out.gz") &&
      (-e $outdir.$refseqFN.".".$restSequence.".l".$length.".mm1.outA.gz") &&
      (-e $outdir.$refseqFN.".".$restSequence.".l".$length.".mm1.outC.gz") &&
      (-e $outdir.$refseqFN.".".$restSequence.".l".$length.".mm1.outG.gz") &&
      (-e $outdir.$refseqFN.".".$restSequence.".l".$length.".mm1.outT.gz") ){
-  print "Refseq files found :\n";
-  print $outdir.$refseqFN.".".$restSequence.".l".$length.".mm0.out.gz"."\n";
-  print $outdir.$refseqFN.".".$restSequence.".l".$length.".mm1.outA.gz"."\n";
-  print $outdir.$refseqFN.".".$restSequence.".l".$length.".mm1.outC.gz"."\n";
-  print $outdir.$refseqFN.".".$restSequence.".l".$length.".mm1.outG.gz"."\n";
-  print $outdir.$refseqFN.".".$restSequence.".l".$length.".mm1.outT.gz"."\n";
+  warn "Refseq files found :\n";
+  warn $outdir.$refseqFN.".".$restSequence.".l".$length.".mm0.out.gz"."\n";
+  warn $outdir.$refseqFN.".".$restSequence.".l".$length.".mm1.outA.gz"."\n";
+  warn $outdir.$refseqFN.".".$restSequence.".l".$length.".mm1.outC.gz"."\n";
+  warn $outdir.$refseqFN.".".$restSequence.".l".$length.".mm1.outG.gz"."\n";
+  warn $outdir.$refseqFN.".".$restSequence.".l".$length.".mm1.outT.gz"."\n";
 }else{
   my $cmd="src/allTag  -l 75 -m 1 -r ".$restSequence." -d $refseq  -o >(sort --temporary-directory=".$outdir." -t \$'\\t' -k3,3 - | gzip -c > ".$outdir.$refseqFN.".".$restSequence.".l".$length.".mm0.out.gz)   -p1 >(sort --temporary-directory=".$outdir." -t \$'\\t' -k3,3 - | gzip -c > ".$outdir.$refseqFN.".".$restSequence.".l".$length.".mm1.outA.gz)   -p2 >(sort --temporary-directory=".$outdir." -t \$'\\t' -k3,3 - | gzip -c > ".$outdir.$refseqFN.".".$restSequence.".l".$length.".mm1.outC.gz)   -p3 >(sort --temporary-directory=".$outdir." -t \$'\\t' -k3,3 - | gzip -c > ".$outdir.$refseqFN.".".$restSequence.".l".$length.".mm1.outG.gz)   -p4 >(sort --temporary-directory=".$outdir." -t \$'\\t' -k3,3 - | gzip -c > ".$outdir.$refseqFN.".".$restSequence.".l".$length.".mm1.outT.gz)";
   systemBash($cmd);
@@ -299,12 +317,12 @@ if ( (-e $outdir.$unigeneFN.".".$restSequence.".l".$length.".mm0.out.gz") &&
      (-e $outdir.$unigeneFN.".".$restSequence.".l".$length.".mm1.outC.gz") &&
      (-e $outdir.$unigeneFN.".".$restSequence.".l".$length.".mm1.outG.gz") &&
      (-e $outdir.$unigeneFN.".".$restSequence.".l".$length.".mm1.outT.gz") ){
-  print "Unigene files found :\n";
-  print $outdir.$unigeneFN.".".$restSequence.".l".$length.".mm0.out.gz"."\n";
-  print $outdir.$unigeneFN.".".$restSequence.".l".$length.".mm1.outA.gz"."\n";
-  print $outdir.$unigeneFN.".".$restSequence.".l".$length.".mm1.outC.gz"."\n";
-  print $outdir.$unigeneFN.".".$restSequence.".l".$length.".mm1.outG.gz"."\n";
-  print $outdir.$unigeneFN.".".$restSequence.".l".$length.".mm1.outT.gz"."\n";
+  warn "Unigene files found :\n";
+  warn $outdir.$unigeneFN.".".$restSequence.".l".$length.".mm0.out.gz"."\n";
+  warn $outdir.$unigeneFN.".".$restSequence.".l".$length.".mm1.outA.gz"."\n";
+  warn $outdir.$unigeneFN.".".$restSequence.".l".$length.".mm1.outC.gz"."\n";
+  warn $outdir.$unigeneFN.".".$restSequence.".l".$length.".mm1.outG.gz"."\n";
+  warn $outdir.$unigeneFN.".".$restSequence.".l".$length.".mm1.outT.gz"."\n";
 }else{
   my $cmd="src/allTag  -l 75 -m 1 -r ".$restSequence." -d $unigene  -o >(sort --temporary-directory=".$outdir." -t \$'\\t' -k3,3 - | gzip -c > ".$outdir.$unigeneFN.".".$restSequence.".l".$length.".mm0.out.gz)   -p1 >(sort --temporary-directory=".$outdir." -t \$'\\t' -k3,3 - | gzip -c > ".$outdir.$unigeneFN.".".$restSequence.".l".$length.".mm1.outA.gz)   -p2 >(sort --temporary-directory=".$outdir." -t \$'\\t' -k3,3 - | gzip -c > ".$outdir.$unigeneFN.".".$restSequence.".l".$length.".mm1.outC.gz)   -p3 >(sort --temporary-directory=".$outdir." -t \$'\\t' -k3,3 - | gzip -c > ".$outdir.$unigeneFN.".".$restSequence.".l".$length.".mm1.outG.gz)   -p4 >(sort --temporary-directory=".$outdir." -t \$'\\t' -k3,3 - | gzip -c > ".$outdir.$unigeneFN.".".$restSequence.".l".$length.".mm1.outT.gz)";
   #print $cmd;
@@ -321,12 +339,12 @@ if ( (-e $outdir.$genomicFN.".".$restSequence.".l".$length.".mm0.out.gz") &&
      (-e $outdir.$genomicFN.".".$restSequence.".l".$length.".mm1.outC.gz") &&
      (-e $outdir.$genomicFN.".".$restSequence.".l".$length.".mm1.outG.gz") &&
      (-e $outdir.$genomicFN.".".$restSequence.".l".$length.".mm1.outT.gz") ){
-  print "Genomic files found :\n";
-  print $outdir.$genomicFN.".".$restSequence.".l".$length.".mm0.out.gz"."\n";
-  print $outdir.$genomicFN.".".$restSequence.".l".$length.".mm1.outA.gz"."\n";
-  print $outdir.$genomicFN.".".$restSequence.".l".$length.".mm1.outC.gz"."\n";
-  print $outdir.$genomicFN.".".$restSequence.".l".$length.".mm1.outG.gz"."\n";
-  print $outdir.$genomicFN.".".$restSequence.".l".$length.".mm1.outT.gz"."\n";
+  warn "Genomic files found :\n";
+  warn $outdir.$genomicFN.".".$restSequence.".l".$length.".mm0.out.gz"."\n";
+  warn $outdir.$genomicFN.".".$restSequence.".l".$length.".mm1.outA.gz"."\n";
+  warn $outdir.$genomicFN.".".$restSequence.".l".$length.".mm1.outC.gz"."\n";
+  warn $outdir.$genomicFN.".".$restSequence.".l".$length.".mm1.outG.gz"."\n";
+  warn $outdir.$genomicFN.".".$restSequence.".l".$length.".mm1.outT.gz"."\n";
 }else{
   my $cmd="src/allTag  -l 75 -m 1 -r ".$restSequence." -d $genomic  -o >(sort --temporary-directory=".$outdir." -t \$'\\t' -k3,3 - | gzip -c > ".$outdir.$genomicFN.".".$restSequence.".l".$length.".mm0.out.gz)   -p1 >(sort --temporary-directory=".$outdir." -t \$'\\t' -k3,3 - | gzip -c > ".$outdir.$genomicFN.".".$restSequence.".l".$length.".mm1.outA.gz)   -p2 >(sort --temporary-directory=".$outdir." -t \$'\\t' -k3,3 - | gzip -c > ".$outdir.$genomicFN.".".$restSequence.".l".$length.".mm1.outC.gz)   -p3 >(sort --temporary-directory=".$outdir." -t \$'\\t' -k3,3 - | gzip -c > ".$outdir.$genomicFN.".".$restSequence.".l".$length.".mm1.outG.gz)   -p4 >(sort --temporary-directory=".$outdir." -t \$'\\t' -k3,3 - | gzip -c > ".$outdir.$genomicFN.".".$restSequence.".l".$length.".mm1.outT.gz)";
   #print $cmd;
@@ -341,7 +359,7 @@ if ( (-e $outdir.$genomicFN.".".$restSequence.".l".$length.".mm0.out.gz") &&
 
 
 if ( (-e $outdir."all".".".$restSequence.".l".$length.".mm0.out") ){
-  print "Found ".$outdir."all".".".$restSequence.".l".$length.".mm0.out\n";
+  warn "Found ".$outdir."all".".".$restSequence.".l".$length.".mm0.out\n";
 }else{
   my $cmd="src/readFiles.pl ".$restSequence." A 0  ".$refseqFN." ".$unigeneFN." ".$genomicFN." $outdir $length  > ".$outdir."all.".$restSequence.".l".$length.".mm0.out";
   #`$cmd`;
@@ -352,7 +370,7 @@ if ( (-e $outdir."all".".".$restSequence.".l".$length.".mm0.out") ){
 }
 
 if ( (-e $outdir."all".".".$restSequence.".l".$length.".mm1.outA") ){
-  print "Found ".$outdir."all".".".$restSequence.".l".$length.".mm1.outA\n";
+  warn "Found ".$outdir."all".".".$restSequence.".l".$length.".mm1.outA\n";
 }else{
   my $cmd="src/readFiles.pl ".$restSequence." A 1 ".$refseqFN." ".$unigeneFN." ".$genomicFN." $outdir $length > ".$outdir."all.".$restSequence.".l".$length.".mm1.outA";
   #`$cmd`;
@@ -364,7 +382,7 @@ if ( (-e $outdir."all".".".$restSequence.".l".$length.".mm1.outA") ){
 }
 
 if ( (-e $outdir."all".".".$restSequence.".l".$length.".mm1.outC") ){
-  print "Found ".$outdir."all".".".$restSequence.".l".$length.".mm1.outC\n";
+  warn "Found ".$outdir."all".".".$restSequence.".l".$length.".mm1.outC\n";
 }else{
   my $cmd="src/readFiles.pl ".$restSequence." C 1 ".$refseqFN." ".$unigeneFN." ".$genomicFN." $outdir $length > ".$outdir."all.".$restSequence.".l".$length.".mm1.outC";
   #`$cmd`;
@@ -376,7 +394,7 @@ if ( (-e $outdir."all".".".$restSequence.".l".$length.".mm1.outC") ){
 }
 
 if ( (-e $outdir."all".".".$restSequence.".l".$length.".mm1.outG") ){
-  print "Found ".$outdir."all".".".$restSequence.".l".$length.".mm1.outG\n";
+  warn "Found ".$outdir."all".".".$restSequence.".l".$length.".mm1.outG\n";
 }else{
   my $cmd="src/readFiles.pl ".$restSequence." G 1 ".$refseqFN." ".$unigeneFN." ".$genomicFN." $outdir $length > ".$outdir."all.".$restSequence.".l".$length.".mm1.outG";
 #  `$cmd`;
@@ -388,7 +406,7 @@ if ( (-e $outdir."all".".".$restSequence.".l".$length.".mm1.outG") ){
 }
 
 if ( (-e $outdir."all".".".$restSequence.".l".$length.".mm1.outT") ){
-  print "Found ".$outdir."all".".".$restSequence.".l".$length.".mm1.outT\n";
+  warn "Found ".$outdir."all".".".$restSequence.".l".$length.".mm1.outT\n";
 }else{
   my $cmd="src/readFiles.pl ".$restSequence." T 1 ".$refseqFN." ".$unigeneFN." ".$genomicFN." $outdir $length > ".$outdir."all.".$restSequence.".l".$length.".mm1.outT";
 #  print $cmd;
@@ -405,7 +423,7 @@ if ( (-e $outdir."all".".".$restSequence.".l".$length.".db.0") ){
   my $dbIndex=0;
 
   while((-e $outdir."all".".".$restSequence.".l".$length.".db.".$dbIndex ) ){
-    print "Found ".$outdir."all".".".$restSequence.".l".$length.".db.".$dbIndex."\n";
+    warn "Found ".$outdir."all".".".$restSequence.".l".$length.".db.".$dbIndex."\n";
     $lastDbIndex=$dbIndex;
     $dbIndex++;
   }
@@ -417,7 +435,7 @@ if ( (-e $outdir."all".".".$restSequence.".l".$length.".db.0") ){
 
   my $dbIndex=0;
   open(FILEOUT,">".$outdir."all.".$restSequence.".l".$length.".db.".$dbIndex) or die "Cannot open ".$outdir."all.".$restSequence.".l".$length.".db.".$dbIndex." for writing\n";
-  print "Writing to ".$outdir."all.".$restSequence.".l".$length.".db.".$dbIndex."\n";
+  warn "Writing to ".$outdir."all.".$restSequence.".l".$length.".db.".$dbIndex."\n";
 
   foreach my $file ($outdir."all.".$restSequence.".l".$length.".mm0.out",
 		    $outdir."all.".$restSequence.".l".$length.".mm1.outA",
@@ -433,7 +451,7 @@ if ( (-e $outdir."all".".".$restSequence.".l".$length.".db.0") ){
 	$dbIndex++;
 	$lastDbIndex=$dbIndex;
 	open(FILEOUT,">".$outdir."all.".$restSequence.".l".$length.".db.".$dbIndex) or die "Cannot open ".$outdir."all.".$restSequence.".l".$length.".db.".$dbIndex." for writing\n";
-	print "Writing to ".$outdir."all.".$restSequence.".l".$length.".db.".$dbIndex."\n";
+	warn "Writing to ".$outdir."all.".$restSequence.".l".$length.".db.".$dbIndex."\n";
       }
     }
     close(FILE);
@@ -444,31 +462,31 @@ if ( (-e $outdir."all".".".$restSequence.".l".$length.".db.0") ){
 
 }
 
-for(my $dbindex=0;$dbindex<=$lastDbIndex;$dbindex++){
-  my $cmd="src/searchtags ".$outdir."all".".".$restSequence.".l".$length.".db.".$dbindex." ".$inputFile." ".$allowedMismatches." > ".$inputFile.".out".$dbindex."\n";
-  #print "Running command= ".$cmd."\n";
-  #`$cmd`;
-  #if($? != 0){
-  #  die "Command ".$cmd." failed";
-  #}
+
+if($onlyCreateDatabase == 0){
+
+
+  for(my $dbindex=0;$dbindex<=$lastDbIndex;$dbindex++){
+    my $cmd="src/searchtags ".$outdir."all".".".$restSequence.".l".$length.".db.".$dbindex." ".$inputFile." ".$allowedMismatches." > ".$inputFile.".out".$dbindex."\n";
+    systemBash($cmd);
+  }
+
+  my @filesToMerge;
+  for(my $dbindex=0;$dbindex<=$lastDbIndex;$dbindex++){
+    push(@filesToMerge,$inputFile.".out".$dbindex);
+  }
+  my $cmd="src/mergeAnnotation.pl ".join(" ",@filesToMerge)." > ".$inputFile.".out";
+
   systemBash($cmd);
 
+
+  for(my $dbindex=0;$dbindex<=$lastDbIndex;$dbindex++){
+    warn "Removing ".$inputFile.".out".$dbindex."\n";
+    unlink($inputFile.".out".$dbindex);
+  }
+
 }
 
-my @filesToMerge;
-for(my $dbindex=0;$dbindex<=$lastDbIndex;$dbindex++){
-  push(@filesToMerge,$inputFile.".out".$dbindex);
-}
-my $cmd="src/mergeAnnotation.pl ".join(" ",@filesToMerge)." > ".$inputFile.".out";
-
-#`$cmd`;
-#if($? != 0){
-#  die "Command ".$cmd." failed";
-#}
-systemBash($cmd);
+warn "Program finished successfully\n";
 
 
-for(my $dbindex=0;$dbindex<=$lastDbIndex;$dbindex++){
-  print "Removing ".$inputFile.".out".$dbindex."\n";
-  unlink($inputFile.".out".$dbindex);
-}
